+++ migration-library/src/test/kotlin/com/quadient/migration/service/deploy/InteractiveDeployClientTest.kt
@@ -82,7 +82,7 @@ class InteractiveDeployClientTest {
         // given
         mockBasicDocumentObjects()
         mockBasicSuccessfulIpsOperations()
-        every { documentObjectBuilder.buildDocumentObject(any()) } returns "<xml />"
+        every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns "<xml />"
         every { documentObjectRepository.findModel(any()) } returns aBlock("99", internal = false)
         every { statusTrackingRepository.findLastEventRelevantToOutput(any(), any(), any()) } returns Active()
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
@@ -98,7 +98,7 @@ class InteractiveDeployClientTest {
                 "icm://Interactive/$tenant/Templates/defaultFolder/0.jld"
             )
         )
-        verify(exactly = 3) { documentObjectBuilder.buildDocumentObject(any()) }
+        verify(exactly = 3) { documentObjectBuilder.buildDocumentObject(any(), any()) }
     }
 
     @Test
@@ -130,7 +130,7 @@ class InteractiveDeployClientTest {
             mockDocumentObject(aTemplate("3", listOf(aDocumentObjectRef("block1"), aDocumentObjectRef("block2"))))
 
         every { documentObjectRepository.list(any()) } returns listOf(externalBlock, template)
-        every { documentObjectBuilder.buildDocumentObject(any()) } returns "<xml />"
+        every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns "<xml />"
         every { documentObjectRepository.findModel(any()) } returns aBlock("99", internal = false)
         every { statusTrackingRepository.findLastEventRelevantToOutput(any(), any(), any()) } returns Active()
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
@@ -147,7 +147,7 @@ class InteractiveDeployClientTest {
                 "icm://Interactive/$tenant/Templates/defaultFolder/3.jld"
             )
         )
-        verify(exactly = 2) { documentObjectBuilder.buildDocumentObject(any()) }
+        verify(exactly = 2) { documentObjectBuilder.buildDocumentObject(any(), any()) }
     }
 
     @Test
@@ -160,7 +160,7 @@ class InteractiveDeployClientTest {
                 any(), any(), any(), any(), "icm://Interactive/$tenant/Blocks/defaultFolder/0.jld"
             )
         } returns OperationResult.Failure("Problem")
-        every { documentObjectBuilder.buildDocumentObject(any()) } returns "<xml />"
+        every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns "<xml />"
         every { documentObjectRepository.findModel(any()) } returns aBlock("99", internal = false)
         every { statusTrackingRepository.findLastEventRelevantToOutput(any(), any(), any()) } returns Active()
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
@@ -180,7 +180,7 @@ class InteractiveDeployClientTest {
             )
         }
         verify { ipsService.close() }
-        verify(exactly = 3) { documentObjectBuilder.buildDocumentObject(any()) }
+        verify(exactly = 3) { documentObjectBuilder.buildDocumentObject(any(), any()) }
     }
 
     @Test
@@ -202,7 +202,7 @@ class InteractiveDeployClientTest {
             mockDocumentObject(aTemplate("3", listOf(aDocumentObjectRef("block1"), aDocumentObjectRef("block2"))))
 
         every { documentObjectRepository.list(any()) } returns listOf(block, template)
-        every { documentObjectBuilder.buildDocumentObject(any()) } returns "<xml />"
+        every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns "<xml />"
         every { documentObjectRepository.findModel(any()) } returns aBlock("99", internal = false)
         every { statusTrackingRepository.findLastEventRelevantToOutput(any(), any(), any()) } returns Active()
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
@@ -219,7 +219,7 @@ class InteractiveDeployClientTest {
                 "icm://Interactive/$tenant/Templates/defaultFolder/3.jld"
             )
         )
-        verify(exactly = 2) { documentObjectBuilder.buildDocumentObject(any()) }
+        verify(exactly = 2) { documentObjectBuilder.buildDocumentObject(any(), any()) }
     }
 
     @Test
@@ -229,7 +229,7 @@ class InteractiveDeployClientTest {
         val block = mockDocumentObject(aBlock(id = "1", listOf(ImageModelRef(image.id))))
 
         every { documentObjectRepository.list(any()) } returns listOf(block)
-        every { documentObjectBuilder.buildDocumentObject(any()) } returns "<xml />"
+        every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns "<xml />"
         every { ipsService.tryUpload(any(), any()) } returns OperationResult.Success
         every { statusTrackingRepository.findLastEventRelevantToOutput(any(), any(), any()) } returns Active()
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
@@ -260,7 +260,7 @@ class InteractiveDeployClientTest {
         )
 
         every { documentObjectRepository.list(any()) } returns listOf(block)
-        every { documentObjectBuilder.buildDocumentObject(any()) } returns "<xml />"
+        every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns "<xml />"
         every { ipsService.upload(any(), any()) } just runs
         every { statusTrackingRepository.findLastEventRelevantToOutput(any(), any(), any()) } returns Active()
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
@@ -292,7 +292,7 @@ class InteractiveDeployClientTest {
 
         every { documentObjectRepository.list(any()) } returns listOf(block)
         every { documentObjectRepository.findModel(innerBlock.id) } returns innerBlock
-        every { documentObjectBuilder.buildDocumentObject(any()) } returns "<xml />"
+        every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns "<xml />"
         every { statusTrackingRepository.findLastEventRelevantToOutput(any(), any(), any()) } returns Active()
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
         every { ipsService.tryUpload(any(), any()) } returns OperationResult.Success
@@ -570,7 +570,7 @@ class InteractiveDeployClientTest {
         every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns aDeployedStatus("id")
         every { documentObjectRepository.findModel(innerBlock.id) } throws IllegalStateException("Not found")
         every { documentObjectRepository.list(any()) } returns listOf(template, block)
-        every { documentObjectBuilder.buildDocumentObject(block) } throws IllegalStateException("Inner block not found")
+        every { documentObjectBuilder.buildDocumentObject(block, any()) } throws IllegalStateException("Inner block not found")
 
         // when
         val result = subject.deployDocumentObjects()
@@ -657,7 +657,7 @@ class InteractiveDeployClientTest {
             val xml = "<xml>${documentObject.nameOrId()}</xml>"
             val outputPath = "icm://${documentObject.nameOrId()}"
 
-            every { documentObjectBuilder.buildDocumentObject(documentObject) } returns xml
+            every { documentObjectBuilder.buildDocumentObject(documentObject, any()) } returns xml
             every { documentObjectBuilder.getDocumentObjectPath(documentObject) } returns outputPath
         }
         return documentObject
@@ -668,7 +668,7 @@ class InteractiveDeployClientTest {
         @BeforeEach
         fun setup() {
             every { ipsService.close() } just runs
-            every { documentObjectBuilder.buildDocumentObject(any()) } returns ""
+            every { documentObjectBuilder.buildDocumentObject(any(), any()) } returns ""
             every { statusTrackingRepository.deployed(any(), any<Uuid>(), any(), any(), any(), any(), any()) } returns  aDeployedStatus("id")
             every { statusTrackingRepository.error(any(), any(), any(), any(), any(), any(), any(), any()) } returns aErrorStatus("id")
             every { statusTrackingRepository.active(any(), any()) } returns aActiveStatus("id")
@@ -697,7 +697,7 @@ class InteractiveDeployClientTest {
             subject.deployDocumentObjectsInternal(docObjects)
 
             // then
-            verify(exactly = 0) { documentObjectBuilder.buildDocumentObject(any()) }
+            verify(exactly = 0) { documentObjectBuilder.buildDocumentObject(any(), any()) }
             verify(exactly = 0) { imageRepository.findModel(any()) }
         }
 
@@ -724,7 +724,7 @@ class InteractiveDeployClientTest {
             subject.deployDocumentObjectsInternal(docObjects)
 
             // then
-            verify(exactly = 2) { documentObjectBuilder.buildDocumentObject(any()) }
+            verify(exactly = 2) { documentObjectBuilder.buildDocumentObject(any(), any()) }
             verify(exactly = 1) { statusTrackingRepository.deployed("D_2", any<Uuid>(), any(), any(), any(), any(), any()) }
             verify(exactly = 1) { statusTrackingRepository.deployed("D_3", any<Uuid>(), any(), any(), any(), any(), any()) }
             verify(exactly = 1) { statusTrackingRepository.deployed("I_1", any<Uuid>(), any(), any(), any(), any(), any()) }
@@ -745,7 +745,7 @@ class InteractiveDeployClientTest {
             subject.deployDocumentObjectsInternal(docObjects)
 
             // then
-            verify(exactly = 1) { documentObjectBuilder.buildDocumentObject(any()) }
+            verify(exactly = 1) { documentObjectBuilder.buildDocumentObject(any(), any()) }
             verify(exactly = 1) { statusTrackingRepository.error("D_1", any(), any(), any(), any(), any(), "oops", any()) }
             verify(exactly = 1) { statusTrackingRepository.error("I_1", any(), any(), any(), any(), any(), any(), any()) }
         }