+++ wfd-xml/api/src/main/java/com/quadient/wfdxml/api/layoutnodes/Root.java
@@ -2,4 +2,5 @@ package com.quadient.wfdxml.api.layoutnodes;
 
 public interface Root {
     Root setAllowRuntimeModifications(boolean allowRuntimeModifications);
+    Root setExternalStylesLayout(String vcsLocation);
 }